Introduce go modules.
This makes the following possible. Suppose you have a project
using Go modules which uses the html2text lib. If you want to
fork html2text in order to create a PR you can then apply the steps outlined below:

- Fork the upstream html2text repo.
- Clone the fork.
- Use the fork in your project by adding a replace directive
  to the go.mod file of your project:

	module gitlab.com/zaunerc/mymodule

	# use my fork until all PRs have been (hopefully) merged upstream.
	replace jaytaylor.com/html2text => ../html2text

	require (
		jaytaylor.com/html2text v0.0.0
	)

If the upstream html2text project does not include a
go.mod file using the replace directive is not supported.